Feature/#179 enable compiler warnings (#230)
* Enable compiler warnings for target build

* Remove compiler warnings in unit tests

* Refactor CmdProcessor unit tests, add const qualifiers where needed in other utils tests

* Disable cast-qual warning for cellular tests, fix a regression in LeanB2CAP tests

* Disable cast-qual warning for the whole cellular module

* Disable some warnings because clang-tidy doesn't recognize them

* Resolved comments for CmdProcessor_unittests
Removed useless cast in tlv

+++ core/essentials/include/mcu/stm32/stm32l4/Kiso_MCU_STM32L4_SPI_Handle.h
@@ -52,7 +52,7 @@ struct MCU_SPI_Transaction_S
 {
     uint8_t *pRxDataBuffer; /**< Reference to the receive buffer used for the transaction*/
     uint8_t *pTxDataBuffer; /**< Reference to the transmit buffer used for the transcation */
-    uint32_t Size; /**< Size in bytes of the transaction */
+    uint16_t Size; /**< Size in bytes of the transaction */
 };
 
 /**
